requirements.txt
pandas
sqlalchemy
psycopg2-binary
openpyxl
jupyter
python-dotenv
